[#17] Make org.jooq.impl.Parameter independent of Field [#17] Cleaned up code

+++ jOOQ-test/src/org/jooq/test/jOOQMySQLTest.java
@@ -75,20 +75,20 @@ import org.junit.Test;
  * @author Lukas Eder
  */
 public class jOOQMySQLTest {
-	
+
 	private Connection connection;
 
 	@Before
 	public void setUp() throws Exception {
 		Configuration.getInstance().setDialect(SQLDialect.MYSQL);
-		
+
 		Class.forName("com.mysql.jdbc.Driver");
 		connection = DriverManager.getConnection ("jdbc:mysql://localhost/test", "root", "");
 
 		Statement stmt = null;
 		File file = new File(getClass().getResource("/org/jooq/test/mysql/create.sql").toURI());
 		String allSQL = FileUtils.readFileToString(file);
-		
+
 		for (String sql : allSQL.split("/")) {
 			try {
 				stmt = connection.createStatement();
@@ -108,61 +108,61 @@ public class jOOQMySQLTest {
 			connection = null;
 		}
 	}
-	
+
 	@Test
 	public final void testSelectSimpleQuery() throws Exception {
 		SelectQuery q = QueryFactory.createSelectQuery();
 		Field<Integer> f1 = Functions.constant(1);
 		Field<Double> f2 = Functions.constant(2d);
 		Field<String> f3 = Functions.constant("test");
-		
+
 		q.addSelect(f1);
 		q.addSelect(f2);
 		q.addSelect(f3);
-		
+
 		int i = q.execute(connection);
 		Result result = q.getResult();
-		
+
 		assertEquals(1, i);
 		assertEquals(1, result.getNumberOfRecords());
 		assertEquals(3, result.getFields().size());
 		assertTrue(result.getFields().contains(f1));
 		assertTrue(result.getFields().contains(f2));
 		assertTrue(result.getFields().contains(f3));
-		
+
 		assertEquals(3, result.getRecords().get(0).getFields().size());
 		assertTrue(result.getRecords().get(0).getFields().contains(f1));
 		assertTrue(result.getRecords().get(0).getFields().contains(f2));
 		assertTrue(result.getRecords().get(0).getFields().contains(f3));
-		
+
 		assertEquals((Integer) 1, result.getRecords().get(0).getValue(f1));
-		assertEquals((Double) 2d, result.getRecords().get(0).getValue(f2));
+		assertEquals(2d, result.getRecords().get(0).getValue(f2));
 		assertEquals("test", result.getRecords().get(0).getValue(f3));
 	}
-	
+
 	@Test
 	public final void testSelectQuery() throws Exception {
 		SelectQuery q = QueryFactory.createSelectQuery(T_AUTHOR);
 		q.addSelect(T_AUTHOR.getFields());
 		q.addOrderBy(LAST_NAME);
-		
+
 		int rows = q.execute(connection);
 		Result result = q.getResult();
-		
+
 		assertEquals(2, rows);
 		assertEquals(2, result.getNumberOfRecords());
 		assertEquals("Coelho", result.getRecord(0).getValue(LAST_NAME));
 		assertEquals("Orwell", result.getRecord(1).getValue(LAST_NAME));
 	}
-	
+
 	@Test
 	public final void testCombinedSelectQuery() throws Exception {
 		SelectQuery q1 = QueryFactory.createSelectQuery(T_BOOK);
 		SelectQuery q2 = QueryFactory.createSelectQuery(T_BOOK);
-		
+
 		q1.addCompareCondition(AUTHOR_ID, 1);
 		q2.addCompareCondition(TITLE, "Brida");
-		
+
 		ResultProviderQuery combine = q1.combine(q2);
 
 		int rows = combine.execute(connection);
@@ -173,20 +173,20 @@ public class jOOQMySQLTest {
 	public final void testJoinQuery() throws Exception {
 		SelectQuery q1 = QueryFactory.createSelectQuery(V_LIBRARY);
 		q1.addOrderBy(VLibrary.TITLE);
-		
+
 		SelectQuery q2 = QueryFactory.createSelectQuery(T_AUTHOR);
 		q2.addJoin(T_BOOK, TBook.AUTHOR_ID, TAuthor.ID);
 		q2.addOrderBy(TBook.TITLE);
-		
+
 		int rows1 = q1.execute(connection);
 		int rows2 = q2.execute(connection);
-		
+
 		assertEquals(4, rows1);
 		assertEquals(4, rows2);
-		
+
 		Result result1 = q1.getResult();
 		Result result2 = q2.getResult();
-		
+
 		assertEquals("1984", result1.getRecord(0).getValue(VLibrary.TITLE));
 		assertEquals("1984", result2.getRecord(0).getValue(TBook.TITLE));
 
@@ -199,59 +199,59 @@ public class jOOQMySQLTest {
 		assertEquals("O Alquimista", result1.getRecord(3).getValue(VLibrary.TITLE));
 		assertEquals("O Alquimista", result2.getRecord(3).getValue(TBook.TITLE));
 	}
-	
+
 	@Test
 	public final void testProcedure() throws Exception {
 		PAuthorExists procedure = new PAuthorExists();
 		procedure.setAuthorName("Paulo");
 		procedure.execute(connection);
 		assertTrue(procedure.getResult());
-		
+
 		procedure = new PAuthorExists();
 		procedure.setAuthorName("Shakespeare");
 		procedure.execute(connection);
 		assertFalse(procedure.getResult());
 	}
-	
+
 	@Test
 	public final void testFunction1() throws Exception {
 		FAuthorExists function1 = new FAuthorExists();
 		function1.setAuthorName("Paulo");
 		function1.execute(connection);
 		assertEquals(1, (int) function1.getReturnValue());
-		
+
 		FAuthorExists function2 = new FAuthorExists();
 		function2.setAuthorName("Shakespeare");
 		function2.execute(connection);
 		assertEquals(0, (int) function2.getReturnValue());
 	}
-	
+
 	@Test
 	public final void testFunction2() throws Exception {
 		// TODO
 		// StoredFunctions cannot be integrated with Functions yet, because
 		// Functions expect fields as parameters, whereas StoredFunctions expect
 		// constant values
-		
-		
+
+
 //		FAuthorExists function1 = new FAuthorExists();
 //		function1.setAuthorName("Paulo");
 //		Function<Byte> f1 = function1.getFunction();
-//		
+//
 //		FAuthorExists function2 = new FAuthorExists();
 //		function2.setAuthorName("Shakespeare");
 //		Function<Byte> f2 = function2.getFunction();
-//		
+//
 //		SelectQuery q = QueryFactory.createSelectQuery();
 //		q.addSelect(f1, f2);
 //		q.execute(connection);
 //		Result result = q.getResult();
-//		
+//
 //		assertEquals(1, result.getNumberOfRecords());
 //		assertEquals(1, (int) result.getRecord(0).getValue(f1));
 //		assertEquals(0, (int) result.getRecord(0).getValue(f2));
 	}
-	
+
 	@Test
 	public final void testFunction3() throws Exception {
 		SelectQuery q1 = QueryFactory.createSelectQuery();
@@ -259,23 +259,23 @@ public class jOOQMySQLTest {
 		Field<Timestamp> ts = now.alias("ts");
 		Field<Date> date = Functions.currentDate().alias("date");
 		Field<Time> time = Functions.currentTime().alias("time");
-		
+
 		Field<Integer> year = Functions.extract(now, DatePart.YEAR).alias("y");
 		Field<Integer> month = Functions.extract(now, DatePart.MONTH).alias("m");
 		Field<Integer> day = Functions.extract(now, DatePart.DAY).alias("day");
 		Field<Integer> hour = Functions.extract(now, DatePart.HOUR).alias("h");
 		Field<Integer> minute = Functions.extract(now, DatePart.MINUTE).alias("mn");
 		Field<Integer> second = Functions.extract(now, DatePart.SECOND).alias("sec");
-		
+
 		q1.addSelect(ts, date, time, year, month, day, hour, minute, second);
 		q1.execute(connection);
-		
+
 		Record record = q1.getResult().getRecord(0);
 		String timestamp = record.getValue(ts).toString();
-		
+
 		assertEquals(timestamp.split(" ")[0], record.getValue(date).toString());
 		assertEquals(timestamp.split(" ")[1], record.getValue(time).toString() + ".0");
-		
+
 		assertEquals(Integer.valueOf(timestamp.split(" ")[0].split("-")[0]), record.getValue(year));
 		assertEquals(Integer.valueOf(timestamp.split(" ")[0].split("-")[1]), record.getValue(month));
 		assertEquals(Integer.valueOf(timestamp.split(" ")[0].split("-")[2]), record.getValue(day));
@@ -283,7 +283,7 @@ public class jOOQMySQLTest {
 		assertEquals(Integer.valueOf(timestamp.split(" ")[1].split(":")[1]), record.getValue(minute));
 		assertEquals(Integer.valueOf(timestamp.split(" ")[1].split(":")[2].split("\\.")[0]), record.getValue(second));
 	}
-	
+
 	@Test
 	public final void testFunction4() throws Exception {
 		SelectQuery q = QueryFactory.createSelectQuery();
@@ -291,31 +291,31 @@ public class jOOQMySQLTest {
 		Field<Integer> charLength = Functions.charLength(constant).alias("len");
 		Field<Integer> bitLength = Functions.bitLength(constant).alias("bitlen");
 		Field<Integer> octetLength = Functions.octetLength(constant).alias("octetlen");
-		
+
 		q.addSelect(charLength, bitLength, octetLength);
 		q.execute(connection);
-		
+
 		Record record = q.getResult().getRecord(0);
-		
+
 		assertEquals((Integer) 3, record.getValue(charLength));
 		assertEquals((Integer) 24, record.getValue(bitLength));
-		assertEquals((Integer) 3, record.getValue(octetLength)); 
+		assertEquals((Integer) 3, record.getValue(octetLength));
 	}
-	
+
 	@Test
 	public final void testFunction5() throws Exception {
 		SelectQuery q = QueryFactory.createSelectQuery(V_LIBRARY);
-		
+
 		Field<String> o = Functions.constant("o");
 		Field<Integer> position = Functions.position(AUTHOR, o).alias("p");
 		q.addSelect(AUTHOR);
 		q.addSelect(position);
 		q.addOrderBy(AUTHOR, ASC);
-		
+
 		q.execute(connection);
 		Record r1 = q.getResult().getRecord(1); // George Orwell
 		Record r2 = q.getResult().getRecord(2); // Paulo Coelho
-		
+
 		assertEquals((Integer) 3, r1.getValue(position));
 		assertEquals((Integer) 5, r2.getValue(position));
 	}
